heuristica andando!!!!!!
Falta cambiar que no sea especifico para esc 1

+++ TP1-AI-arriola-pernuzzi/src/frsf/cidisi/exercise/noinformadaprofundidad/search/actions/IdentificarVictimario.java
@@ -51,7 +51,7 @@ public class IdentificarVictimario extends SearchAction {
      */
     @Override
     public Double getCost() {
-        return new Double(0);
+        return new Double(5);
     }
 
     /**
